services/StreamingServer/requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
requests
watchdog
pyjwt
